Reorganize project to access all tools via `import arc`
