+++ .gitignore
@@ -2,7 +2,7 @@
 .pytest_cache/
 *.egg-info/
 **/*.pyc
-*/__pycache__
+**/__pycache__
 build/
 dist/
 MANIFEST
