+++ arc/events/__init__.py
@@ -1,16 +1,17 @@
 import json
 import urllib.request
 import boto3
-
-from arc import services
-from arc.lib import use_aws, get_ports
+from arc.services import _services
+from arc._lib import use_aws, get_ports
 
 port = None
 cache = {}
 
 
 def parse(event):
-    messages = list(map((lambda record: json.loads(record["body"])), event["Records"]))
+    messages = list(
+        map((lambda record: json.loads(record["Sns"]["Message"])), event["Records"])
+    )
     if len(messages) == 1:
         return messages[0]
     return messages
@@ -24,25 +25,23 @@ def publish(name, payload):
         try:
             dump = json.dumps({"name": name, "payload": payload})
             data = bytes(dump.encode())
-            handler = urllib.request.urlopen(f"http://localhost:{port}/queues", data)
+            handler = urllib.request.urlopen(f"http://localhost:{port}/events", data)
             return handler.read().decode("utf-8")
         except Exception as error:
-            print("arc.queues.publish to Sandbox failed: " + str(error))
+            print("arc.events.publish to Sandbox failed: " + str(error))
             return data
 
     def publish_aws(name, payload):
         global cache
 
         def pub(arn):
-            sqs = boto3.client("sqs")
-            return sqs.send_message(
-                QueueUrl=arn, MessageBody=json.dumps(payload), DelaySeconds=0
-            )
+            sns = boto3.client("sns")
+            return sns.publish(TopicArn=arn, Message=json.dumps(payload))
 
         if cache.get(name):
             return pub(cache[name])
-        service_map = services()
-        cache = service_map["queues"]
+        service_map = _services()
+        cache = service_map["events"]
         arn = cache[name]
         if not arn:
             raise TypeError(f"{name} event not found")
@@ -53,7 +52,7 @@ def publish(name, payload):
     if local:
         ports = get_ports()
         if not ports.get("events"):
-            raise TypeError("Sandbox queues port not found")
+            raise TypeError("Sandbox events port not found")
         port = ports["events"]
         return publish_sandbox(name, payload)
     return publish_aws(name, payload)
